Fix eslint/tsconfig for Windows
On Windows, npm explore eslint-plugin-jolly-roger seems to result in the
pwd being node_modules/eslint-plugin-jolly-roger, instead of eslint
(like we see on POSIX platforms), which causes the extend to break.

There's no philosophical motivation for having that inheritance in the
first place - the two projects are fairly disconnected and I only used
the "extends" the first time out of laziness, but this at least gets the
Windows install a little further along.