Make SSH user list configurable in CloudFormation config.
We add a parameter to specify a CSV of the users and their SSH keys, and
we use a lambda function to convert it into the "users:" stanza for the
EC2 config.

An alternative could be to specify two parallel parameters, one for the
usernames, and one for the SSH keys. We could then use the
AWS::LanguageExtensions function Fn::ForEach to loop over one list while
using Fn::Select to pull the corresponding object from the other list.
This would simplify the implementation, but make the "API" (the
parameters) arguably more complex.

Fixes #2013