Add support for signing in with a Google Account.
Show a "Sign in with Google" button on the login page. This sign in flow
first attempts to match against existing users who have linked that same
Google Account. If there is no such user, or multiple such users, the
request is rejected.

We use a custom login handler rather than the standard loginWithGoogle
module for simplicity since we already have existing linked users that
would be difficult to migrate to the standard system due to not storing
the necessary credentials. Since we have enough to validate identity in
our current schema, this is fine for now.

Icon taken from:
https://developers.google.com/identity/branding-guidelines

Notes / issues:

- This fixes a bug in the existing "unlink Google Account" flow where
the linked email is cleared but not the linked account ID.

- At least on local instances, a console warning appears upon showing
the Google Account popup: "Cross-Origin-Opener-Policy policy would block
the window.closed call". This also occurs for the existing Google
Account integration so isn't exclusive to this flow.

- Sometimes, when you click Sign Out, the login page renders with a
blank "Sign in with Google" button, though the button appears to work as
expected.

See #2021